@@ -23,6 +23,7 @@ import info.fmro.shared.stream.definitions.OrderChangeMessage;
 import info.fmro.shared.stream.objects.PoisonPill;
 import info.fmro.shared.stream.objects.RunnerId;
 import info.fmro.shared.stream.objects.SerializableObjectModification;
+import info.fmro.shared.stream.objects.StreamHasStartedPill;
 import info.fmro.shared.stream.objects.StreamObjectInterface;
 import info.fmro.shared.stream.objects.StreamSynchronizedMap;
 import info.fmro.shared.stream.protocol.ChangeMessageFactory;
@@ -68,24 +69,22 @@ public class SSLClientThread
 
     @SuppressWarnings({"OverlyComplexMethod", "OverlyLongMethod", "OverlyNestedMethod", "SwitchStatementDensity", "unchecked", "ChainOfInstanceofChecks", "OverlyCoupledMethod"})
     private synchronized void runAfterReceive(@NotNull final StreamObjectInterface receivedCommand) {
-        if (receivedCommand instanceof RulesManager) {
-            @NotNull final RulesManager rulesManager = (RulesManager) receivedCommand;
+        if (receivedCommand instanceof @NotNull final StreamHasStartedPill streamHasStartedPill) {
+            Statics.clearCachedObjects();
+//            GUI.publicRefreshDisplayedManagedObject();
+        } else if (receivedCommand instanceof @NotNull final RulesManager rulesManager) {
             Statics.rulesManager.copyFromStream(rulesManager);
             GUI.publicRefreshDisplayedManagedObject();
-        } else if (receivedCommand instanceof ExistingFunds) {
-            @NotNull final ExistingFunds existingFunds = (ExistingFunds) receivedCommand;
+        } else if (receivedCommand instanceof @NotNull final ExistingFunds existingFunds) {
             Statics.existingFunds.copyFrom(existingFunds);
-            GUI.publicRefreshDisplayedManagedObject();
-        } else if (receivedCommand instanceof MarketCache) {
-            @NotNull final MarketCache marketCache = (MarketCache) receivedCommand;
+//            GUI.publicRefreshDisplayedManagedObject();
+        } else if (receivedCommand instanceof @NotNull final MarketCache marketCache) {
             SharedStatics.marketCache.copyFromStream(marketCache);
-            GUI.publicRefreshDisplayedManagedObject();
-        } else if (receivedCommand instanceof OrderCache) {
-            @NotNull final OrderCache orderCache = (OrderCache) receivedCommand;
+            GUI.publicRefreshDisplayedManagedObject(true);
+        } else if (receivedCommand instanceof @NotNull final OrderCache orderCache) {
             SharedStatics.orderCache.copyFromStream(orderCache);
-            GUI.publicRefreshDisplayedManagedObject();
-        } else if (receivedCommand instanceof StreamSynchronizedMap<?, ?>) {
-            @NotNull final StreamSynchronizedMap<?, ?> streamSynchronizedMap = (StreamSynchronizedMap<?, ?>) receivedCommand;
+            GUI.publicRefreshDisplayedManagedObject(true);
+        } else if (receivedCommand instanceof @NotNull final StreamSynchronizedMap<?, ?> streamSynchronizedMap) {
             final Class<?> clazz = streamSynchronizedMap.getClazz();
             if (MarketCatalogue.class.equals(clazz)) {
                 @SuppressWarnings("unchecked") @NotNull final StreamSynchronizedMap<String, MarketCatalogue> streamMarketCatalogueMap = (StreamSynchronizedMap<String, MarketCatalogue>) receivedCommand;
@@ -98,8 +97,7 @@ public class SSLClientThread
             } else {
                 logger.error("unknown streamSynchronizedMap class in runAfterReceive for: {} {}", clazz, Generic.objectToString(receivedCommand));
             }
-        } else if (receivedCommand instanceof MarketChangeMessage) {
-            @NotNull final MarketChangeMessage marketChangeMessage = (MarketChangeMessage) receivedCommand;
+        } else if (receivedCommand instanceof @NotNull final MarketChangeMessage marketChangeMessage) {
             SharedStatics.marketCache.onMarketChange(ChangeMessageFactory.ToChangeMessage(-1, marketChangeMessage), Statics.rulesManager, Statics.marketCataloguesMap);
             final HashSet<String> marketIds = marketChangeMessage.getChangedMarketIds();
             if (marketIds == null) { // normal, nothing to be done
@@ -110,8 +108,7 @@ public class SSLClientThread
                     GUI.managedEventUpdated(eventId);
                 }
             }
-        } else if (receivedCommand instanceof OrderChangeMessage) {
-            @NotNull final OrderChangeMessage orderChangeMessage = (OrderChangeMessage) receivedCommand;
+        } else if (receivedCommand instanceof @NotNull final OrderChangeMessage orderChangeMessage) {
             SharedStatics.orderCache.onOrderChange(ChangeMessageFactory.ToChangeMessage(-1, orderChangeMessage), Statics.rulesManager);
             final HashSet<String> marketIds = orderChangeMessage.getChangedMarketIds();
             if (marketIds == null) { // normal, nothing to be done
@@ -122,18 +119,14 @@ public class SSLClientThread
                     GUI.managedEventUpdated(eventId);
                 }
             }
-        } else if (receivedCommand instanceof SerializableObjectModification) {
-            final SerializableObjectModification<?> serializableObjectModification = (SerializableObjectModification<?>) receivedCommand;
+        } else if (receivedCommand instanceof final SerializableObjectModification<?> serializableObjectModification) {
             final Enum<?> command = serializableObjectModification.getCommand();
-            if (command instanceof RulesManagerModificationCommand) {
-                final RulesManagerModificationCommand rulesManagerModificationCommand = (RulesManagerModificationCommand) command;
+            if (command instanceof final RulesManagerModificationCommand rulesManagerModificationCommand) {
                 final Object[] objectsToModify = serializableObjectModification.getArray();
                 switch (rulesManagerModificationCommand) {
                     case addManagedEvent:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof ManagedEvent) {
-                                final String eventId = (String) objectsToModify[0];
-                                final ManagedEvent managedEvent = (ManagedEvent) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String eventId && objectsToModify[1] instanceof final ManagedEvent managedEvent) {
                                 Statics.rulesManager.addManagedEvent(eventId, managedEvent);
                                 GUI.managedEventUpdated(eventId);
                             } else {
@@ -145,8 +138,7 @@ public class SSLClientThread
                         break;
                     case removeManagedEvent:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof String) {
-                                final String eventId = (String) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final String eventId) {
                                 Statics.rulesManager.removeManagedEvent(eventId, Statics.marketCataloguesMap);
                             } else {
                                 logger.error("wrong objectsToModify class in runAfterReceive: {} {} {}", Generic.objectToString(objectsToModify), rulesManagerModificationCommand.name(), Generic.objectToString(receivedCommand));
@@ -157,9 +149,7 @@ public class SSLClientThread
                         break;
                     case addManagedMarket:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof ManagedMarket) {
-                                final String marketId = (String) objectsToModify[0];
-                                final ManagedMarket managedMarket = (ManagedMarket) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final ManagedMarket managedMarket) {
                                 Statics.rulesManager.addManagedMarket(marketId, managedMarket);
                                 GUI.managedMarketUpdated(marketId);
                             } else {
@@ -171,8 +161,7 @@ public class SSLClientThread
                         break;
                     case removeManagedMarket:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof String) {
-                                final String marketId = (String) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final String marketId) {
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
                                 Statics.rulesManager.removeManagedMarket(marketId, Statics.marketCataloguesMap);
                                 GUI.managedEventUpdated(eventId);
@@ -185,9 +174,7 @@ public class SSLClientThread
                         break;
                     case setEventAmountLimit:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof Double) {
-                                final String eventId = (String) objectsToModify[0];
-                                final Double newAmount = (Double) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String eventId && objectsToModify[1] instanceof final Double newAmount) {
                                 Statics.rulesManager.setEventAmountLimit(eventId, newAmount, Statics.existingFunds, Statics.marketCataloguesMap);
                                 GUI.managedEventUpdated(eventId);
                             } else {
@@ -199,8 +186,7 @@ public class SSLClientThread
                         break;
                     case addManagedRunner:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof ManagedRunner) {
-                                final ManagedRunner managedRunner = (ManagedRunner) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final ManagedRunner managedRunner) {
                                 final String marketId = managedRunner.getMarketId();
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
                                 Statics.rulesManager.addManagedRunner(managedRunner, Statics.marketCataloguesMap, Statics.eventsMap);
@@ -215,10 +201,8 @@ public class SSLClientThread
                         break;
                     case removeManagedRunner:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof RunnerId) {
-                                final String marketId = (String) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final RunnerId runnerId) {
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
-                                final RunnerId runnerId = (RunnerId) objectsToModify[1];
                                 Statics.rulesManager.removeManagedRunner(marketId, runnerId);
                                 GUI.managedMarketUpdated(marketId);
                                 GUI.managedEventUpdated(eventId);
@@ -231,10 +215,8 @@ public class SSLClientThread
                         break;
                     case setMarketAmountLimit:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof Double) {
-                                final String marketId = (String) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final Double amountLimit) {
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
-                                final Double amountLimit = (Double) objectsToModify[1];
                                 Statics.rulesManager.setMarketAmountLimit(marketId, amountLimit, Statics.existingFunds);
                                 GUI.managedMarketUpdated(marketId);
                                 GUI.managedEventUpdated(eventId);
@@ -247,10 +229,7 @@ public class SSLClientThread
                         break;
                     case setBackAmountLimit:
                         if (objectsToModify != null && objectsToModify.length == 3) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof RunnerId && objectsToModify[2] instanceof Double) {
-                                final String marketId = (String) objectsToModify[0];
-                                final RunnerId runnerId = (RunnerId) objectsToModify[1];
-                                final Double amountLimit = (Double) objectsToModify[2];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final RunnerId runnerId && objectsToModify[2] instanceof final Double amountLimit) {
                                 Statics.rulesManager.setRunnerBackAmountLimit(marketId, runnerId, amountLimit);
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
                                 GUI.managedMarketUpdated(marketId);
@@ -264,10 +243,7 @@ public class SSLClientThread
                         break;
                     case setLayAmountLimit:
                         if (objectsToModify != null && objectsToModify.length == 3) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof RunnerId && objectsToModify[2] instanceof Double) {
-                                final String marketId = (String) objectsToModify[0];
-                                final RunnerId runnerId = (RunnerId) objectsToModify[1];
-                                final Double amountLimit = (Double) objectsToModify[2];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final RunnerId runnerId && objectsToModify[2] instanceof final Double amountLimit) {
                                 Statics.rulesManager.setRunnerLayAmountLimit(marketId, runnerId, amountLimit);
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
                                 GUI.managedMarketUpdated(marketId);
@@ -281,10 +257,7 @@ public class SSLClientThread
                         break;
                     case setMinBackOdds:
                         if (objectsToModify != null && objectsToModify.length == 3) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof RunnerId && objectsToModify[2] instanceof Double) {
-                                final String marketId = (String) objectsToModify[0];
-                                final RunnerId runnerId = (RunnerId) objectsToModify[1];
-                                final Double odds = (Double) objectsToModify[2];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final RunnerId runnerId && objectsToModify[2] instanceof final Double odds) {
                                 Statics.rulesManager.setRunnerMinBackOdds(marketId, runnerId, odds);
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
                                 GUI.managedMarketUpdated(marketId);
@@ -298,10 +271,7 @@ public class SSLClientThread
                         break;
                     case setMaxLayOdds:
                         if (objectsToModify != null && objectsToModify.length == 3) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof RunnerId && objectsToModify[2] instanceof Double) {
-                                final String marketId = (String) objectsToModify[0];
-                                final RunnerId runnerId = (RunnerId) objectsToModify[1];
-                                final Double odds = (Double) objectsToModify[2];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final RunnerId runnerId && objectsToModify[2] instanceof final Double odds) {
                                 Statics.rulesManager.setRunnerMaxLayOdds(marketId, runnerId, odds);
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
                                 GUI.managedMarketUpdated(marketId);
@@ -315,9 +285,7 @@ public class SSLClientThread
                         break;
                     case setMarketName:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof String) {
-                                final String marketId = (String) objectsToModify[0];
-                                final String marketName = (String) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final String marketName) {
                                 Statics.rulesManager.setMarketName(marketId, marketName);
                                 GUI.managedMarketUpdated(marketId);
                             } else {
@@ -329,9 +297,7 @@ public class SSLClientThread
                         break;
                     case setEventName:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof String) {
-                                final String eventId = (String) objectsToModify[0];
-                                final String eventName = (String) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String eventId && objectsToModify[1] instanceof final String eventName) {
                                 Statics.rulesManager.setEventName(eventId, eventName);
                                 GUI.managedEventUpdated(eventId);
                             } else {
@@ -343,9 +309,7 @@ public class SSLClientThread
                         break;
                     case setMarketEnabled:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof Boolean) {
-                                final String marketId = (String) objectsToModify[0];
-                                final Boolean marketEnabled = (Boolean) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final Boolean marketEnabled) {
                                 Statics.rulesManager.setMarketEnabled(marketId, marketEnabled);
                                 GUI.managedMarketUpdated(marketId);
                             } else {
@@ -357,9 +321,7 @@ public class SSLClientThread
                         break;
                     case setMarketMandatoryPlace:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof Boolean) {
-                                final String marketId = (String) objectsToModify[0];
-                                final Boolean mandatoryPlace = (Boolean) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final Boolean mandatoryPlace) {
                                 Statics.rulesManager.setMarketMandatoryPlace(marketId, mandatoryPlace);
                                 GUI.managedMarketUpdated(marketId);
                             } else {
@@ -371,10 +333,7 @@ public class SSLClientThread
                         break;
                     case setRunnerMandatoryPlace:
                         if (objectsToModify != null && objectsToModify.length == 3) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof RunnerId && objectsToModify[2] instanceof Boolean) {
-                                final String marketId = (String) objectsToModify[0];
-                                final RunnerId runnerId = (RunnerId) objectsToModify[1];
-                                final Boolean mandatoryPlace = (Boolean) objectsToModify[2];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final RunnerId runnerId && objectsToModify[2] instanceof final Boolean mandatoryPlace) {
                                 Statics.rulesManager.setRunnerMandatoryPlace(marketId, runnerId, mandatoryPlace);
                                 GUI.managedMarketUpdated(marketId);
                             } else {
@@ -386,9 +345,7 @@ public class SSLClientThread
                         break;
                     case setMarketKeepAtInPlay:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof Boolean) {
-                                final String marketId = (String) objectsToModify[0];
-                                final Boolean keepAtInPlay = (Boolean) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final Boolean keepAtInPlay) {
                                 Statics.rulesManager.setMarketKeepAtInPlay(marketId, keepAtInPlay);
                                 GUI.managedMarketUpdated(marketId);
                             } else {
@@ -400,10 +357,7 @@ public class SSLClientThread
                         break;
                     case setRunnerPrefSide:
                         if (objectsToModify != null && objectsToModify.length == 3) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof RunnerId && objectsToModify[2] instanceof PrefSide) {
-                                final String marketId = (String) objectsToModify[0];
-                                final RunnerId runnerId = (RunnerId) objectsToModify[1];
-                                final PrefSide prefSide = (PrefSide) objectsToModify[2];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final RunnerId runnerId && objectsToModify[2] instanceof final PrefSide prefSide) {
                                 Statics.rulesManager.setRunnerPrefSide(marketId, runnerId, prefSide);
                                 GUI.managedMarketUpdated(marketId);
                             } else {
@@ -415,9 +369,7 @@ public class SSLClientThread
                         break;
                     case setMarketCalculatedLimit:
                         if (objectsToModify != null && objectsToModify.length == 2) {
-                            if (objectsToModify[0] instanceof String && objectsToModify[1] instanceof Double) {
-                                final String marketId = (String) objectsToModify[0];
-                                final Double calculatedLimit = (Double) objectsToModify[1];
+                            if (objectsToModify[0] instanceof final String marketId && objectsToModify[1] instanceof final Double calculatedLimit) {
                                 Statics.rulesManager.setMarketCalculatedLimit(marketId, calculatedLimit, Statics.existingFunds);
                                 GUI.managedMarketUpdated(marketId);
                             } else {
@@ -429,8 +381,7 @@ public class SSLClientThread
                         break;
                     case addTempOrder:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof TemporaryOrder) {
-                                @NotNull final TemporaryOrder temporaryOrder = (TemporaryOrder) objectsToModify[0];
+                            if (objectsToModify[0] instanceof @NotNull final TemporaryOrder temporaryOrder) {
 
                                 final String marketId = temporaryOrder.getMarketId();
                                 final RunnerId runnerId = temporaryOrder.getRunnerId();
@@ -454,8 +405,7 @@ public class SSLClientThread
                         break;
                     case removeTempOrder:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof TemporaryOrder) {
-                                @NotNull final TemporaryOrder temporaryOrder = (TemporaryOrder) objectsToModify[0];
+                            if (objectsToModify[0] instanceof @NotNull final TemporaryOrder temporaryOrder) {
                                 SharedStatics.orderCache.removeTempOrder(temporaryOrder);
                                 final String marketId = temporaryOrder.getMarketId();
                                 final String eventId = Formulas.getEventIdOfMarketId(marketId, Statics.marketCataloguesMap);
@@ -471,14 +421,12 @@ public class SSLClientThread
                     default:
                         logger.error("unknown rulesManagerModificationCommand in runAfterReceive: {} {}", rulesManagerModificationCommand.name(), Generic.objectToString(receivedCommand));
                 } // end switch
-            } else if (command instanceof ExistingFundsModificationCommand) {
-                final ExistingFundsModificationCommand existingFundsModificationCommand = (ExistingFundsModificationCommand) command;
+            } else if (command instanceof final ExistingFundsModificationCommand existingFundsModificationCommand) {
                 final Object[] objectsToModify = serializableObjectModification.getArray();
                 switch (existingFundsModificationCommand) {
                     case setCurrencyRate:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof Double) {
-                                final Double rate = (Double) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final Double rate) {
                                 Statics.existingFunds.setCurrencyRate(rate);
                                 GUI.publicRefreshDisplayedManagedObject();
                             } else {
@@ -490,8 +438,7 @@ public class SSLClientThread
                         break;
                     case setReserve:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof Double) {
-                                final Double reserve = (Double) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final Double reserve) {
                                 Statics.existingFunds.setReserve(reserve);
                                 GUI.publicRefreshDisplayedManagedObject();
                             } else {
@@ -503,8 +450,7 @@ public class SSLClientThread
                         break;
                     case setAvailableFunds:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof Double) {
-                                final Double availableFunds = (Double) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final Double availableFunds) {
                                 Statics.existingFunds.setAvailableFunds(availableFunds);
                             } else {
                                 logger.error("wrong objectsToModify class in runAfterReceive: {} {} {}", Generic.objectToString(objectsToModify), existingFundsModificationCommand.name(), Generic.objectToString(receivedCommand));
@@ -515,8 +461,7 @@ public class SSLClientThread
                         break;
                     case setExposure:
                         if (objectsToModify != null && objectsToModify.length == 1) {
-                            if (objectsToModify[0] instanceof Double) {
-                                final Double exposure = (Double) objectsToModify[0];
+                            if (objectsToModify[0] instanceof final Double exposure) {
                                 Statics.existingFunds.setExposure(exposure);
                             } else {
                                 logger.error("wrong objectsToModify class in runAfterReceive: {} {} {}", Generic.objectToString(objectsToModify), existingFundsModificationCommand.name(), Generic.objectToString(receivedCommand));
@@ -528,13 +473,11 @@ public class SSLClientThread
                     default:
                         logger.error("unknown existingFundsModificationCommand in runAfterReceive: {} {}", existingFundsModificationCommand.name(), Generic.objectToString(receivedCommand));
                 } // end switch
-            } else if (command instanceof SynchronizedMapModificationCommand) {
-                @NotNull final SynchronizedMapModificationCommand synchronizedMapModificationCommand = (SynchronizedMapModificationCommand) command;
+            } else if (command instanceof @NotNull final SynchronizedMapModificationCommand synchronizedMapModificationCommand) {
                 final Object[] objectsToModify = serializableObjectModification.getArray();
-                if (objectsToModify == null || !(objectsToModify[0] instanceof Class<?>)) {
+                if (objectsToModify == null || !(objectsToModify[0] instanceof final Class<?> clazz)) {
                     logger.error("improper objectsToModify for SynchronizedMapModificationCommand: {} {} {}", Generic.objectToString(objectsToModify), synchronizedMapModificationCommand.name(), Generic.objectToString(receivedCommand));
                 } else {
-                    final Class<?> clazz = (Class<?>) objectsToModify[0];
                     @Nullable final ClientStreamSynchronizedMap<String, Serializable> mapToUse;
 
                     if (MarketCatalogue.class.equals(clazz)) {
@@ -561,8 +504,7 @@ public class SSLClientThread
                                 break;
                             case put:
                                 if (nObjects >= 3) {
-                                    if (objectsToModify[1] instanceof String && Generic.objectInstanceOf(clazz, objectsToModify[2])) {
-                                        final String key = (String) objectsToModify[1];
+                                    if (objectsToModify[1] instanceof final String key && Generic.objectInstanceOf(clazz, objectsToModify[2])) {
                                         final Serializable value = (Serializable) objectsToModify[2];
                                         if (nObjects == 3) {
                                             mapToUse.put(key, value);
@@ -583,8 +525,7 @@ public class SSLClientThread
                                 break;
                             case putIfAbsent:
                                 if (nObjects == 3) {
-                                    if (objectsToModify[1] instanceof String && Generic.objectInstanceOf(clazz, objectsToModify[2])) {
-                                        final String key = (String) objectsToModify[1];
+                                    if (objectsToModify[1] instanceof final String key && Generic.objectInstanceOf(clazz, objectsToModify[2])) {
                                         final Serializable value = (Serializable) objectsToModify[2];
                                         mapToUse.putIfAbsent(key, value);
                                         info.fmro.client.utility.Utils.checkMapValues(mapToUse, Set.of(key));
@@ -610,8 +551,7 @@ public class SSLClientThread
                                 break;
                             case remove:
                                 if (nObjects >= 2) {
-                                    if (objectsToModify[1] instanceof String) {
-                                        final String key = (String) objectsToModify[1];
+                                    if (objectsToModify[1] instanceof final String key) {
                                         if (nObjects == 2) {
                                             mapToUse.remove(key);
                                         } else if (nObjects == 3 && Generic.objectInstanceOf(clazz, objectsToModify[2])) {
@@ -772,14 +712,13 @@ public class SSLClientThread
                         nErrors = 0;
                         writerThread = new SSLWriterThread(this.socket, this.sendQueue);
                         writerThread.start();
-                        //noinspection resource,IOResourceOpenedButNotSafelyClosed
+
                         objectInputStream = new ObjectInputStream(this.socket.getInputStream());
 
                         Object receivedObject;
                         do {
                             receivedObject = objectInputStream.readObject();
-                            if (receivedObject instanceof StreamObjectInterface) {
-                                final StreamObjectInterface receivedCommand = (StreamObjectInterface) receivedObject;
+                            if (receivedObject instanceof final StreamObjectInterface receivedCommand) {
 
                                 runAfterReceive(receivedCommand);
                             } else if (receivedObject == null) { // nothing to be done, will reach end of loop and exit loop
@@ -790,12 +729,6 @@ public class SSLClientThread
                     } else {
                         logger.error("STRANGE socket null in SSLClientThread thread, timeStamp={}", System.currentTimeMillis());
                     }
-//                } catch (EOFException e) {
-//                    if (Statics.mustStop.get()) {
-//                        logger.info("EOFException received in SSLClientThread while program stops");
-//                    } else {
-//                        logger.error("EOFException received in SSLClientThread: ", e);
-//                    }
                 } catch (@SuppressWarnings("OverlyBroadCatchBlock") IOException e) {
                     if (SharedStatics.mustStop.get()) {
                         logger.info("IOException received in SSLClientThread, as program stops");
